requirements.txt
requests==2.22.0
Pillow==6.2.1
instabot==0.71.0
python-dotenv==0.10.3
